using more semantic html and less div
Replaced several more div-elements with semantically more meaningful
tags. This in turn reduced the amount of classes used so far. I also
removed some divs which were not really needed. Reduced width and
height definitions in CSS. (The diff on index.html appears so massive
due to the removed mainBG-div element and moving the lines one tab to
the left.)